move 'evolve' tests to separate file
